requirements.txt
# Flight Price Tracker - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Redis & Caching
redis==5.0.1

# Task Queue
celery==5.3.6

# HTTP Client
httpx==0.26.0

# Telegram Bot
aiogram==3.4.1
python-telegram-bot==21.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
